+++ Dynamic_Programming/Resolves/minimising_coins.cpp
@@ -1,13 +1,13 @@
 #include "bits/stdc++.h"
 using namespace std;
 #define debug(x) cout << #x << " = " << x << endl;
-const int inf = 1e9+7;
+const int inf = (int)1e9 + 7;
 
 int main() {
     int n, x;
     cin >> n >> x;
-    vector <int> coins(n);
-    vector <int> dp(x+1);
+    vector<int> coins(n);
+    vector<int> dp(x + 1);
     dp[0] = 0;
     for (int i = 0; i < n; i++) {
         scanf("%d", &coins[i]);
@@ -16,7 +16,7 @@ int main() {
         dp[i] = inf;
         for (auto c : coins) {
             if (i - c >= 0) {
-                dp[i] = min(dp[i-c] + 1, dp[i]);
+                dp[i] = min(dp[i - c] + 1, dp[i]);
             }
         }
     }
